backend/requirements.txt
torch==2.2.0
torchvision==0.17.0
Flask
flask_cors
pillow
timm==0.4.*
scipy
